api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "api"
authors = ["balas <balashenkoaleksandr294@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
drf-spectacular = "^0.27.2"
dj-database-url = "^2.1.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-filter = "^24.2"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.4"
black = "^24.4.2"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "django.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "rest_framework.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "django_filters.*"
ignore_missing_imports = true

[tool.ruff]
lint.select = ["E", "F", "ICN", "TID", "N", "D", "ANN", "S", "RUF"]
lint.ignore = ["D107", "D104", "D106", "D101", "ANN101", "D401", "D407", "D212", "D205", "ANN205", "D202", "D203", "D406",
    "ANN204", "ANN202", "ANN201"]
lint.fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ",
    "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE",
    "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
lint.unfixable = []
line-length = 88
exclude = ["migrations", "venv"]
